requirements.txt
pymongo
python-dotenv
fastapi
uvicorn
sentence-transformers
numpy
